+++ src/prelude.rs
@@ -1,0 +1,10 @@
+pub use crate::debug::{
+    DebugPickingPlugin,
+};
+pub use highlight::{
+    HighlightablePickMesh,
+    PickHighlightParams,
+};
+pub use select::{
+    SelectablePickMesh,
+};
